backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
azure-identity>=1.19,<2.0
openai>=1.60,<2.0
httpx>=0.28,<1.0
pydantic>=2.10,<3.0
gridstatus>=0.29,<1.0
gridstatusio>=0.2,<1.0
pandas>=2.2,<3.0
